# Use zlib_into.unshuffle to improve performance on decompression